tranfers js to typscript
