--- src/components/Grid/Col.js
@@ -1,7 +1,0 @@
-import React from "react";
-
-export const Col =({ size, children }) => (
-    <div className={size.split(" ").map(size => "col-" + size).join(" ")}>
-        {children}
-    </div>
-)
