--- src/components/Grid/Container.js
@@ -1,7 +1,0 @@
-import React from "react";
-
-export const Container =({ fluid, children }) => (
-    <div className={`container${fluid ? "-fluid" : ""}`}>
-        {children}
-    </div>
-)
