--- src/components/Grid/Row.js
@@ -1,7 +1,0 @@
-import React from "react";
-
-export const Row = ({fluid, children }) => (
-    <div className={`row${fluid ? "-fluid" : ""}`}>
-        {children}
-    </div>
-)
