+++ src/components/Results/Results.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import "./Results.css";
 
-const Results = props => (
+const Results = (props: { children: React.ReactNode; }) => (
     <ul className="list-group" id="itemGroup">
         <li className="list-group-item" id="item">
             {props.children}
