+++ src/components/index.tsx
@@ -1,3 +1,2 @@
-export * from "./Grid";
 export * from "./Input";
 export * from "./Results";
